+++ src/main/java/com/first/flash/climbing/problem/infrastructure/ProblemRepositoryImpl.java
@@ -3,8 +3,8 @@ package com.first.flash.climbing.problem.infrastructure;
 import com.first.flash.climbing.problem.domain.Problem;
 import com.first.flash.climbing.problem.domain.ProblemRepository;
 import com.first.flash.climbing.problem.domain.QueryProblem;
-import com.first.flash.climbing.problem.infrastructure.dto.Cursor;
-import com.first.flash.climbing.problem.infrastructure.dto.SortBy;
+import com.first.flash.climbing.problem.infrastructure.paging.Cursor;
+import com.first.flash.climbing.problem.infrastructure.paging.SortBy;
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
